product-market-fit/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
anthropic==0.18.1
openai==1.12.0
langgraph==0.0.25
langchain==0.1.6
langchain-anthropic==0.1.4
langchain-openai==0.0.5
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiosqlite==0.19.0
python-multipart==0.0.6
tavily-python==0.3.3
